+++ src/services/template/bot.service.ts
@@ -1,34 +1,30 @@
-import { Message } from "discord.js";
+import { CacheType, CommandInteraction } from "discord.js";
 import { injectable, singleton } from "tsyringe";
 import AddTemplateCommand from "../../commands/template/add.command";
 import DeleteTemplateCommand from "../../commands/template/delete.command";
 import GetTemplateCommand from "../../commands/template/get.command";
 import ListTemplateCommand from "../../commands/template/list.command";
-import { IService } from "../../interfaces/service.interface";
+import { ISlashCommand } from "../../interfaces/slash.command.interface";
+import { SlashService } from "../../interfaces/slash.service.interface";
 
 @injectable()
 @singleton()
-export default class TemplateService extends IService<Message> {
+export default class TemplateService extends SlashService {
   public serviceName: string = "template";
-  public serviceDescription: string = "Template Manager Service";
+  public serviceDescription: string = "Template Service";
 
   constructor(
     addCommand: AddTemplateCommand,
+    deleteCommand: DeleteTemplateCommand,
     getCommand: GetTemplateCommand,
-    listCommand: ListTemplateCommand,
-    deleteCommand: DeleteTemplateCommand
+    listCommand: ListTemplateCommand
   ) {
-    super(addCommand, getCommand, listCommand, deleteCommand);
+    super(addCommand, deleteCommand, getCommand, listCommand);
+    this.initializeSlashBuilder(Array.from(arguments) as ISlashCommand[]);
   }
 
-  public async process(message: Message, args: string[]) {
-    if (args.length <= 1) {
-      return this.printServiceInfo(message);
-    }
-    const commandName = args[1];
-    if (!this.serviceContainer.hasOwnProperty(commandName)) {
-      return message.reply("Command not found. Please try again");
-    }
-    await this.serviceContainer[commandName].execute(message, args);
+  public process(interaction: CommandInteraction<CacheType>) {
+    const subcommand = interaction.options.getSubcommand();
+    return this.serviceContainer[subcommand].execute(interaction);
   }
 }
